rename transfer offers method
