+++ amadeus/shopping/__init__.py
@@ -4,9 +4,9 @@ from ._flight_offers_search import FlightOffersSearch
 from ._hotel_offer_search import HotelOfferSearch
 from ._hotel_offers_search import HotelOffersSearch
 from ._activities import Activities
-from ._transfer_offers_search import TransferOffersSearch
+from ._transfer_offers import TransferOffersSearch
 
 __all__ = ['FlightDates', 'FlightDestinations',
            'HotelOffersSearch', 'HotelOfferSearch',
            'FlightOffersSearch', 'Activities',
-           'TransferOffersSearch']
+           'TransferOffers']
