+++ specs/namespaces/test_namespaces.py
@@ -67,7 +67,7 @@ def test_expected_paths(client):
     assert client.duty_of_care is not None
     assert client.duty_of_care.diseases.covid19_report is not None
     assert client.airline.destinations is not None
-    assert client.shopping.transfer_offers_search is not None
+    assert client.shopping.transfer_offers is not None
     assert client.ordering.transfer_orders is not None
     assert client.ordering.transfer_order is not None
 
@@ -548,8 +548,8 @@ def test_airline_destinations_get(client_setup):
     )
 
 
-def test_shopping_transfer_offers_search_post(client_setup):
-    client_setup.shopping.transfer_offers_search.post({'foo': 'bar'})
+def test_shopping_transfer_offers_post(client_setup):
+    client_setup.shopping.transfer_offers.post({'foo': 'bar'})
     client_setup.post.assert_called_with(
         '/v1/shopping/transfer-offers', {'foo': 'bar'}
     )
